Make more compatibility improvements

+++ libc/nt/process.h
@@ -58,9 +58,8 @@ uint32_t GetEnvironmentVariable(const char16_t *lpName, char16_t *lpBuffer,
                                 uint32_t nSize);
 uint32_t SetEnvironmentVariable(const char16_t *lpName, char16_t *lpValue);
 int32_t SetEnvironmentStrings(char16_t *NewEnvironment);
-bool32 GetProcessAffinityMask(int64_t hProcess,
-                              uintptr_t *lpProcessAffinityMask,
-                              uintptr_t *lpSystemAffinityMask);
+bool32 GetProcessAffinityMask(int64_t hProcess, uint64_t *lpProcessAffinityMask,
+                              uint64_t *lpSystemAffinityMask);
 uint64_t /*bool32*/ SetProcessAffinityMask(int64_t hProcess,
                                            uintptr_t dwProcessAffinityMask);
